pyproject.toml
[project]
name = "ai-sdk-python"
packages = ["src/ai_sdk"]
version = "0.1.0"
description = "The easiest, safest, and cleanest way to use AI in your Python code. Inspired by the Vercel AI SDK."
readme = "README.md"
authors = [
    { name = "Asanshay Gupta", email = "asanshay@bharatgupta.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.2",
    "python-dotenv>=1.1.1",
    "pydantic>=2.7.1",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "python-dotenv>=1.1.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/ai_sdk"]
